requirements.txt
altair==5.4.1
base58==2.1.1
blinker==1.8.2
certifi==2024.8.30
chart-studio==1.1.0
colorama==0.4.6
commonmark==0.9.1
decorator==5.1.1
entrypoints==0.4
executing==2.1.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
importlib-metadata==6.6.0
ipykernel==6.29.5
ipython==8.27.0
jsonschema==4.23.0
kaleido==0.2.1
matplotlib==3.8.0
matplotlib-inline==0.1.7
mdurl==0.1.2
numpy>=1.21,<2
Pillow==10.4.0
plotly==5.24.1
pandas==2.2.2
psutil==6.0.0
pydeck==0.9.1
pyarrow==18.0.0
pyrsistent==0.19.3
pytz==2024.2
requests==2.32.3
rich==13.9.3
smmap==5.0.1
stack-data==0.6.3
streamlit==1.40.1
tenacity==9.0.0
toml==0.10.2
tornado==6.4.1
typing-extensions==4.12.2
urllib3==2.2.3
watchdog==5.0.3
